refactor: Auth related handlers moved to auth package rather than spread into user handlers
