+++ internal/auth/auth_middleware/auth_middleware.go
@@ -1,0 +1,33 @@
+package auth_middleware
+
+import (
+	"net/http"
+	"strings"
+
+	"github.com/csusmGDSC/csusmgdsc-api/config"
+	"github.com/csusmGDSC/csusmgdsc-api/internal/auth/auth_utils"
+	"github.com/labstack/echo/v4"
+)
+
+func AuthMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
+	return func(c echo.Context) error {
+		authHeader := c.Request().Header.Get("Authorization")
+		if authHeader == "" {
+			return echo.NewHTTPError(http.StatusUnauthorized, "missing authorization header")
+		}
+
+		cfg := config.LoadConfig()
+		// Extract token from "Bearer <token>"
+		tokenString := strings.TrimPrefix(authHeader, "Bearer ")
+		claims, err := auth_utils.ValidateJWT(tokenString, []byte(cfg.JWTAccessSecret))
+		if err != nil {
+			return echo.NewHTTPError(http.StatusUnauthorized, "invalid token")
+		}
+
+		// Add user info to context
+		c.Set("user_id", claims.UserID)
+		c.Set("user_role", claims.Role)
+
+		return next(c)
+	}
+}
