+++ internal/auth/auth_models/refresh_token.go
@@ -1,4 +1,4 @@
-package models
+package auth_models
 
 import (
 	"time"
