+++ internal/auth/auth_repositories/user_repo.go
@@ -1,4 +1,4 @@
-package repositories
+package auth_repositories
 
 import (
 	"database/sql"
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/csusmGDSC/csusmgdsc-api/internal/auth/auth_models"
 	"github.com/csusmGDSC/csusmgdsc-api/internal/models"
 	"github.com/google/uuid"
 	"github.com/lib/pq"
@@ -22,23 +23,15 @@ func NewUserRepository(db *sql.DB) *UserRepository {
 func (r *UserRepository) Create(user *models.User) error {
 	query := `
 		INSERT INTO users (
-			id, full_name, first_name, last_name, email, password, 
-			role, position, branch, graduation_date,
-			created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12)
+			id, email, password, provider, created_at, updated_at
+		) VALUES ($1, $2, $3, $4, $5, $6)
 	`
 
 	_, err := r.db.Exec(query,
 		user.ID,
-		user.FullName,
-		user.FirstName,
-		user.LastName,
 		user.Email,
 		user.Password,
-		user.Role,
-		user.Position,
-		user.Branch,
-		user.GraduationDate,
+		user.Provider,
 		user.CreatedAt,
 		user.UpdatedAt,
 	)
@@ -48,7 +41,7 @@ func (r *UserRepository) Create(user *models.User) error {
 
 // Fields that are specified as NOT NULL in the database must be assigned
 // from current user data before updating
-func (r *UserRepository) Update(userID string, req models.UpdateUserRequest) (*models.User, error) {
+func (r *UserRepository) Update(userID string, req auth_models.UpdateUserRequest) (*models.User, error) {
 	userUuid, err := uuid.Parse(userID)
 	if err != nil {
 		return nil, err
@@ -60,35 +53,53 @@ func (r *UserRepository) Update(userID string, req models.UpdateUserRequest) (*m
 		return nil, err
 	}
 
-	// Fullname has to be updated if firstName or lastName are provided
-	firstName := currentUser.FirstName
+	// If currentUser.FirstName, currentUser.LastName, or currentUser.FullName are nil,
+	// set to empty strings, this may occur if the user has not been onboarded yet
+	// or onboarded incorrectly
+	firstName := ""
+	lastName := ""
+	fullName := ""
+
+	// User has already been onboarded, so use current values
+	if currentUser.FirstName != nil {
+		firstName = *currentUser.FirstName
+	}
+	if currentUser.LastName != nil {
+		lastName = *currentUser.LastName
+	}
+	if currentUser.FullName != nil {
+		fullName = *currentUser.FullName
+	}
+
 	if req.FirstName != nil {
 		firstName = *req.FirstName
 	}
-
-	lastName := currentUser.LastName
 	if req.LastName != nil {
 		lastName = *req.LastName
 	}
-
-	fullName := currentUser.FullName
 	if req.FirstName != nil || req.LastName != nil {
 		fullName = strings.TrimSpace(firstName + " " + lastName)
 	}
 
-	position := currentUser.Position
+	// If position and branch have not been set, default to student and project branch
+	position := models.Student
+	branch := models.Projects
+	if currentUser.Position != nil {
+		position = *currentUser.Position
+	}
+	if currentUser.Branch != nil {
+		branch = *currentUser.Branch
+	}
 	if req.Position != nil {
 		position = *req.Position
 	}
-
-	branch := currentUser.Branch
 	if req.Branch != nil {
 		branch = *req.Branch
 	}
 
 	query := `
-        UPDATE users 
-        SET 
+        UPDATE users
+        SET
             full_name = $1,
             first_name = $2,
             last_name = $3,
@@ -104,7 +115,7 @@ func (r *UserRepository) Update(userID string, req models.UpdateUserRequest) (*m
             graduation_date = $13,
             updated_at = $14
         WHERE id = $15
-        RETURNING id, full_name, first_name, last_name, email, image, 
+        RETURNING id, full_name, first_name, last_name, email, image,
 				role, position, branch, github, linkedin,
                 instagram, discord, bio, tags, website, graduation_date,
                 created_at, updated_at
@@ -170,7 +181,7 @@ func (r *UserRepository) GetByEmail(email string) (*models.User, error) {
 		SELECT id, full_name, first_name, last_name, email, password,
 		 	role, position, branch, image, github,
 			linkedin, instagram, discord, bio, tags, website,
-			graduation_date, created_at, updated_at
+			graduation_date, created_at, updated_at, provider, auth_id
 		FROM users
 		WHERE email = $1
 	`
@@ -196,6 +207,8 @@ func (r *UserRepository) GetByEmail(email string) (*models.User, error) {
 		&user.GraduationDate,
 		&user.CreatedAt,
 		&user.UpdatedAt,
+		&user.Provider,
+		&user.AuthID,
 	)
 
 	if err != nil {
