+++ internal/auth/auth_utils/oauth.go
@@ -1,55 +1,18 @@
-package auth
+package auth_utils
 
 import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"sync"
 
-	"github.com/csusmGDSC/csusmgdsc-api/config"
+	"github.com/csusmGDSC/csusmgdsc-api/internal/auth"
 	"golang.org/x/oauth2"
-	"golang.org/x/oauth2/github"
-	"golang.org/x/oauth2/google"
 )
-
-var (
-	once         sync.Once
-	githubConfig *oauth2.Config
-	googleConfig *oauth2.Config
-)
-
-type OAuthUserData struct {
-	ID        string  `json:"id"`
-	Name      string  `json:"name"`
-	Email     *string `json:"email,omitempty"`
-	AvatarURL *string `json:"avatar_url,omitempty"`
-	Provider  string  `json:"provider"`
-}
-
-func InitOAuth() {
-	once.Do(func() {
-		cfg := config.LoadConfig()
-
-		githubConfig = &oauth2.Config{
-			ClientID:     cfg.GitHubClientID,
-			ClientSecret: cfg.GitHubClientSecret,
-			Endpoint:     github.Endpoint,
-			RedirectURL:  cfg.OAuthRedirectUrl + "/auth/github/callback",
-			Scopes:       []string{"user:email"},
-		}
-
-		googleConfig = &oauth2.Config{
-			ClientID:     cfg.GoogleClientID,
-			ClientSecret: cfg.GoogleClientSecret,
-			Endpoint:     google.Endpoint,
-			RedirectURL:  cfg.OAuthRedirectUrl + "/auth/google/callback",
-			Scopes:       []string{"email", "profile"},
-		}
-	})
-}
 
 // GetOAuthURL generates an OAuth URL for the provider
 func GetOAuthURL(provider string) (string, error) {
+	githubConfig := auth.GetGitHubConfig()
+	googleConfig := auth.GetGoogleConfig()
 	switch provider {
 	case "github":
 		if githubConfig == nil {
@@ -67,7 +30,9 @@ func GetOAuthURL(provider string) (string, error) {
 }
 
 // HandleOAuthCallback handles OAuth2 callback and returns user data
-func HandleOAuthCallback(provider string, code string) (*OAuthUserData, error) {
+func HandleOAuthCallback(provider string, code string) (*auth.OAuthUserData, error) {
+	githubConfig := auth.GetGitHubConfig()
+	googleConfig := auth.GetGoogleConfig()
 	var config *oauth2.Config
 	var userInfoURL string
 
@@ -104,7 +69,7 @@ func HandleOAuthCallback(provider string, code string) (*OAuthUserData, error) {
 		return nil, fmt.Errorf("failed to decode user info: %v", err)
 	}
 
-	userData := &OAuthUserData{
+	userData := &auth.OAuthUserData{
 		Provider: provider,
 	}
