+++ routes/routes.go
@@ -1,27 +1,25 @@
 package routes
 
 import (
-	"github.com/csusmGDSC/csusmgdsc-api/internal/auth"
+	"github.com/csusmGDSC/csusmgdsc-api/internal/auth/auth_handlers"
+	"github.com/csusmGDSC/csusmgdsc-api/internal/auth/auth_middleware"
 	"github.com/csusmGDSC/csusmgdsc-api/internal/handlers"
 	"github.com/labstack/echo/v4"
 )
 
 func InitRoutes(e *echo.Echo, h *handlers.Handler) {
 	e.POST("/events", h.InsertEventHandler)
+}
 
+func InitOAuthRoutes(e *echo.Echo, h *auth_handlers.OAuthHandler) {
 	authGroup := e.Group("/auth")
 	authGroup.POST("/register", h.RegisterUser)
 	authGroup.POST("/login", h.LoginUser)
 	authGroup.PATCH("/refresh", h.RefreshUser)
 	authGroup.POST("/logout", h.LogoutUser)
-	authGroup.POST("/logoutAll", h.LogoutAll, auth.AuthMiddleware)
-	authGroup.GET("/:provider/login", auth.OAuthLogin)
-	authGroup.GET("/:provider/callback", auth.OAuthCallback)
-	authGroup.POST("/complete-registration", auth.CompleteOAuthRegistration)
-
-	// User routes can be used called from User or Admin
-	apiUserGroup := e.Group("/user")
-	apiUserGroup.Use(auth.AuthMiddleware)
-	apiUserGroup.PUT("/update/:id", h.UpdateUser)
-	apiUserGroup.DELETE("/delete/:id", h.DeleteUser)
+	authGroup.POST("/logoutAll", h.LogoutAll, auth_middleware.AuthMiddleware)
+	authGroup.GET("/:provider/login", h.OAuthLogin)
+	authGroup.GET("/:provider/callback", h.OAuthCallback)
+	authGroup.PUT("/update/:id", h.UpdateUser, auth_middleware.AuthMiddleware)
+	authGroup.DELETE("/delete/:id", h.DeleteUser, auth_middleware.AuthMiddleware)
 }
